requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
stripe==7.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
python-json-logger==2.0.7
whitenoise==6.6.0
Pillow==10.1.0
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
PyJWT==2.8.0
cryptography==41.0.5
pynacl==1.5.0